DEBUG: デザイン作業用: yarn dev でダミー番組・ダミーコメント表示

+++ app/services/nicolive-program/nicolive-comment-filter.ts
@@ -67,6 +67,9 @@ export class NicoliveCommentFilterService extends StatefulService<INicoliveComme
   }
 
   async fetchFilters() {
+    if (process.env.DEV_SERVER) {
+      return;
+    }
     const result = await this.client.fetchFilters(this.programID);
     if (!isOk(result)) {
       throw NicoliveFailure.fromClientError('fetchFilters', result);
@@ -76,6 +79,9 @@ export class NicoliveCommentFilterService extends StatefulService<INicoliveComme
   }
 
   async addFilter(record: Omit<FilterRecord, 'id'>) {
+    if (process.env.DEV_SERVER) {
+      return;
+    }
     const result = await this.client.addFilters(this.programID, [record]);
     if (!isOk(result)) {
       throw NicoliveFailure.fromClientError('addFilters', result);
@@ -94,6 +100,9 @@ export class NicoliveCommentFilterService extends StatefulService<INicoliveComme
   }
 
   async deleteFilters(ids: number[]) {
+    if (process.env.DEV_SERVER) {
+      return;
+    }
     const result = await this.client.deleteFilters(this.programID, ids);
     if (!isOk(result)) {
       throw NicoliveFailure.fromClientError('deleteFilters', result);
